RayTracingMaster Script
OnRenderImage() is called by Unity as soon as the camera has finished rendering the meshes onto the screen.

After this, the shader is Dispatched; this means that I am telling the GPU to get busy with a number of thread groups. The size and number of these thread groups can be set in 3 dimensions, which makes it easy to apply these compute shaders. The default thread group size in the unity template is (8, 8, 1), meaning one thread is spawned per 8x8 pixels. Finally the result of this is drawn on the screen using Graphics.Blit()